+++ src/main/java/com/eventify/platform/planning/domain/model/queries/GetAllQuotesByOrganizerIdQuery.java
@@ -1,4 +1,4 @@
 package com.eventify.platform.planning.domain.model.queries;
 
-public record GetAllQuotesByOrganizerIdQuery(int organizerId) {
+public record GetAllQuotesByOrganizerIdQuery(Long organizerId) {
 }
